refactor: Move initializations for presets to root vue app
