+++ src/App.vue
@@ -11,6 +11,12 @@ import { Component, Vue } from 'vue-property-decorator';
 import AppHeader from './components/AppHeader/AppHeader.vue';
 import StreamEditor from './components/StreamEditor/StreamEditor.vue';
 import BottomNav from './components/BottomNav/BottomNav.vue';
+import { StreamDataset } from './core/StreamDataset';
+import {
+  domainStreamEditorModule,
+  domainStreamColorizerModule,
+} from './store/modules/internal';
+import { ColorDefinition } from './core/ColorDefinition';
 
 @Component({
   components: {
@@ -19,7 +25,79 @@ import BottomNav from './components/BottomNav/BottomNav.vue';
     BottomNav,
   },
 })
-export default class App extends Vue {}
+export default class App extends Vue.extend({
+  computed: {
+    ...domainStreamColorizerModule.mapState(['colorMatcherSourceCode']),
+    ...domainStreamColorizerModule.mapGetters(['colorDefinitions']),
+  },
+  methods: {
+    ...domainStreamEditorModule.mapMutations(['pushStreamDataset']),
+    ...domainStreamColorizerModule.mapMutations([
+      'setColorMatcherSourceCode',
+      'setColorDefinitions',
+    ]),
+  },
+}) {
+  public initializeStreamDatasets() {
+    const streamDatasets: StreamDataset[] = [
+      new StreamDataset({
+        sourceCode: "fromEvent(document.defaultView, 'click')",
+      }),
+      new StreamDataset({
+        sourceCode: '_0$.pipe(buffer(_0$.pipe(debounceTime(250))))',
+      }),
+      new StreamDataset({
+        sourceCode: '_1$.pipe(map(ev => ev.length))',
+      }),
+      new StreamDataset({
+        sourceCode: '_2$.pipe(filter(ev => ev === 2))',
+      }),
+    ];
+    streamDatasets.forEach(streamDataset => {
+      this.pushStreamDataset({ streamDataset });
+    });
+  }
+
+  public initializeColorDefinitions() {
+    if (this.colorDefinitions.length) {
+      return;
+    }
+    const blankDefinitions = new Array(9 * 9 - 4)
+      .fill(void 0)
+      .map(() => new ColorDefinition());
+    const definitions = [
+      new ColorDefinition({ colorCode: '#FC5137' }),
+      new ColorDefinition({ colorCode: '#72D329' }),
+      new ColorDefinition({ colorCode: '#FEC137' }),
+      new ColorDefinition({ colorCode: '#348FBF' }),
+    ].concat(blankDefinitions);
+    this.setColorDefinitions(definitions);
+  }
+
+  public initializeColorMatcherSourceCode() {
+    if (this.colorMatcherSourceCode.length) {
+      return;
+    }
+    this.setColorMatcherSourceCode(
+      `
+event => {
+  switch (event) {
+    case 2:
+      return 'a1';
+    default:
+      return null;
+  }
+}
+      `.trim(),
+    );
+  }
+
+  public created() {
+    this.initializeStreamDatasets();
+    this.initializeColorDefinitions();
+    this.initializeColorMatcherSourceCode();
+  }
+}
 </script>
 <style lang="stylus">
 *
